+++ src/entities/auth.ts
@@ -1,4 +1,6 @@
-export interface IAuthData {
+import { FirebaseAuthTypes } from '@react-native-firebase/auth';
+
+export interface IAuthData extends FirebaseAuthTypes.User {
     nome?: string;
     success?: boolean;
     token?: string;